提升hiz精度

+++ src/main/java/rogo/sketchrender/culling/HizTarget.java
@@ -23,7 +23,7 @@ public class HizTarget extends RenderTarget {
         GlStateManager._texParameter(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MIN_FILTER, GL11.GL_NEAREST);
         GlStateManager._texParameter(GL11.GL_TEXTURE_2D, GL11.GL_TEXTURE_MAG_FILTER, GL11.GL_NEAREST);
         GlStateManager._texParameter(GL11.GL_TEXTURE_2D, GL30.GL_GENERATE_MIPMAP, GL11.GL_FALSE);
-        GlStateManager._texImage2D(GL11.GL_TEXTURE_2D, 0, GL30.GL_R8, this.width, this.height, 0, GL11.GL_RED, GL11.GL_UNSIGNED_BYTE, null);
+        GlStateManager._texImage2D(GL11.GL_TEXTURE_2D, 0, GL30.GL_R16F, this.width, this.height, 0, GL11.GL_RED, GL11.GL_UNSIGNED_BYTE, null);
         GlStateManager._glBindFramebuffer(GL30.GL_FRAMEBUFFER, this.frameBufferId);
         GlStateManager._glFramebufferTexture2D(GL30.GL_FRAMEBUFFER, GL30.GL_COLOR_ATTACHMENT0, GL11.GL_TEXTURE_2D, this.colorTextureId, 0);
     }
